requirements.txt
# Core dependencies
anthropic>=0.18.0
boto3>=1.34.0
pandas>=2.0.0
numpy>=1.24.0

# ML/Data Science
scikit-learn>=1.3.0
matplotlib>=3.7.0
seaborn>=0.12.0
scipy>=1.11.0
plotly>=5.18.0

# Jupyter
jupyter>=1.0.0
jupyterlab>=4.0.0
ipywidgets>=8.1.0

# Security tool integrations
crowdstrike-falconpy>=1.3.0
azure-identity>=1.15.0
msgraph-sdk>=1.1.0
msgraph-core>=1.0.0
azure-mgmt-security>=5.0.0
requests>=2.31.0

# MCP protocol
mcp>=0.9.0

# Network analysis
networkx>=3.1
scapy>=2.5.0

# Utilities
python-dateutil>=2.8.0
tqdm>=4.66.0
joblib>=1.3.0

# Async support
aiohttp>=3.9.0

# Data validation
pydantic>=2.0.0
